test-requirements.txt
selenium==4.10.0
webdriver-manager==3.8.6
requests==2.31.0
python-dotenv==1.0.0
supabase==2.3.4
beautifulsoup4==4.12.2
lxml==4.9.3
